Исправлены ошибки роутов, валидация айди
