+++ app.js
@@ -9,7 +9,7 @@ const helmet = require('helmet');
 const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 
-const { PORT = 3000 } = process.env;
+const { PORT } = process.env;
 const cors = require('cors');
 
 const options = {
@@ -27,11 +27,11 @@ const options = {
 
 const mongoose = require('mongoose');
 const { errors } = require('celebrate');
+const signupRoute = require('./routes/signup');
+const signinRoute = require('./routes/signin');
 const userRoutes = require('./routes/users');
 const movieRoutes = require('./routes/movies');
-const { signUpValidation, signInValidation } = require('./middlewares/validation');
 const { auth } = require('./middlewares/auth');
-const { createUser, login } = require('./controllers/users');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 
 const NotFoundError = require('./errors/not-found-error');
@@ -45,6 +45,8 @@ const limiter = rateLimit({
   max: 100,
 });
 
+app.use(requestLogger);
+
 app.use(limiter);
 
 app.use('*', cors(options));
@@ -54,12 +56,10 @@ app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(helmet());
 
-app.use(requestLogger);
-
 app.use(cookieParser());
 
-app.post('/signup', signUpValidation, createUser);
-app.post('/signin', signInValidation, login);
+app.use(signupRoute);
+app.use(signinRoute);
 app.post('/signout', (req, res) => {
   res
     .clearCookie('jwt')
@@ -77,17 +77,6 @@ app.use((req, res, next) => {
 app.use(errorLogger);
 
 app.use(errors());
-
-app.use((err, req, res, next) => {
-  const { statusCode = 500, message } = err;
-
-  res.status(statusCode).send({
-    message: statusCode === 500
-      ? 'На сервере произошла ошибка'
-      : message,
-  });
-  next();
-});
 
 app.use((err, req, res, next) => {
   const { statusCode = 500, message } = err;
