+++ controllers/signin.js
@@ -1,0 +1,44 @@
+const bcrypt = require('bcryptjs');
+
+const jwt = require('jsonwebtoken');
+const User = require('../models/user');
+
+const AuthError = require('../errors/auth-error');
+
+const { NODE_ENV, JWT_SECRET } = process.env;
+
+// авторизация пользователя
+const login = (req, res, next) => {
+  const { email, password } = req.body;
+  User.findOne({ email }).select('+password')
+    .then((user) => {
+      if (!user) {
+        return Promise.reject(new Error('Неправильные почта или пароль'));
+      }
+      return bcrypt.compare(password, user.password)
+        .then((matched) => {
+          if (!matched) {
+            return Promise.reject(new Error('Неправильные почта или пароль'));
+          }
+          return user;
+        });
+    })
+    .then((user) => {
+      const token = jwt.sign(
+        { _id: user._id },
+        NODE_ENV === 'production' ? JWT_SECRET : 'dev-secret',
+        { expiresIn: '7d' },
+      );
+      res.cookie('jwt', token, {
+        maxAge: 3600000 * 24 * 7,
+        httpOnly: true,
+
+      }).sendStatus(200);
+    })
+    .catch((err) => {
+      next(new AuthError(err.message));
+    })
+    .catch(next);
+};
+
+module.exports = { login };
