+++ models/movie.js
@@ -71,7 +71,7 @@ const movieSchema = new mongoose.Schema({
   },
 
   movieId: {
-    type: String,
+    type: Number,
     required: true,
   },
